Extract SearchFab
